# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RAKT KAVACH - National Digital Blood Grid API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Patients list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create patient profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/requests:
    get:
      operationId: getPatientRequests
      tags: [patients]
      summary: Get patient blood requests
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequestListResponse"

  # HOSPITALS
...
  /wallet/{userId}:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get wallet for user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get wallet transactions
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionListResponse"
  /wallet/transfer:
    post:
      operationId: walletTransfer
      tags: [wallet]
      summary: Transfer credits between family members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransferInput"
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"

  # NOTIFICATIONS
...
components:
  schemas:
...
    PatientInput:
      type: object
      required: [bloodGroup, dateOfBirth, gender]
      properties:
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        weight:
          type: number
        height:
          type: number
        abhaId:
          type: string
        medicalNotes:
          type: string
        address:
          type: string
        state:
          type: string
        district:
          type: string
    PatientUpdate:
      type: object
      properties:
        weight:
          type: number
        height:
          type: number
        medicalNotes:
          type: string
        address:
          type: string
    Patient:
      type: object
      required: [id, userId, bloodGroup, dateOfBirth, gender, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        bloodGroup:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        abhaId:
          type: ["string", "null"]
        medicalNotes:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        createdAt:
          type: string
    PatientListResponse:
      type: object
      required: [patients, total, page, limit]
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # HOSPITAL SCHEMAS
...
    BloodRequest:
      type: object
      required: [id, bloodGroup, units, urgency, status, createdAt]
      properties:
        id:
          type: integer
        bloodGroup:
          type: string
        units:
          type: number
        urgency:
          type: string
        status:
          type: string
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        hospitalId:
          type: integer
        hospitalName:
          type: ["string", "null"]
        bloodBankId:
          type: ["integer", "null"]
        bloodBankName:
          type: ["string", "null"]
        requiredBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BloodRequestListResponse:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/BloodRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # EMERGENCY SCHEMAS
...
    Wallet:
      type: object
      required: [id, userId, bloodCredits, donationCredits, familyProtected, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        bloodCredits:
          type: number
        donationCredits:
          type: number
        emergencyCredits:
          type: number
        familyProtected:
          type: boolean
        qrCode:
          type: ["string", "null"]
        createdAt:
          type: string
    WalletTransaction:
      type: object
      required: [id, walletId, type, amount, description, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        type:
          type: string
          enum: [CREDIT_DONATION, DEBIT_EMERGENCY, INTRA_FAMILY_TRANSFER, BONUS, PENALTY]
        amount:
          type: number
        description:
          type: string
        referenceId:
          type: ["string", "null"]
        balanceAfter:
          type: number
        createdAt:
          type: string
    WalletTransactionListResponse:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    WalletTransferInput:
      type: object
      required: [fromUserId, toUserId, amount]
      properties:
        fromUserId:
          type: integer
        toUserId:
          type: integer
        amount:
          type: number
        reason:
          type: string
    TransferResult:
      type: object
      required: [success, message, fromBalance, toBalance]
      properties:
        success:
          type: boolean
        message:
          type: string
        fromBalance:
          type: number
        toBalance:
          type: number

    # NOTIFICATION SCHEMAS
...